int -> long

+++ src/main/java/com/mitchej123/hodgepodge/mixins/early/minecraft/MixinChunk_FixInvalidEntity.java
@@ -50,8 +50,8 @@ public class MixinChunk_FixInvalidEntity {
             return;
         }
 
-        int entityChunkX = MathHelper.floor_double(entity.posX / 16.0D);
-        int entityChunkY = MathHelper.floor_double(entity.posZ / 16.0D);
+        long entityChunkX = MathHelper.floor_double(entity.posX / 16.0D);
+        long entityChunkY = MathHelper.floor_double(entity.posZ / 16.0D);
         long dist = Math.abs(entityChunkX - this.xPosition) + Math.abs(entityChunkY - this.zPosition);
 
         // If it's more than 1000 chunks away from its expected position, it's safe to assume it's invalid
